Externalize startup functions
